Adjusted the UI layout of the button group beside the file browser
