--- lib/features/file/page.dart
@@ -1,41 +1,0 @@
-/// File management page for HealthPod.
-///
-/// Copyright (C) 2025, Software Innovation Institute, ANU
-///
-/// Licensed under the GNU General Public License, Version 3 (the "License");
-///
-/// License: https://www.gnu.org/licenses/gpl-3.0.en.html
-//
-// This program is free software: you can redistribute it and/or modify it under
-// the terms of the GNU General Public License as published by the Free Software
-// Foundation, either version 3 of the License, or (at your option) any later
-// version.
-//
-// This program is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
-// details.
-//
-// You should have received a copy of the GNU General Public License along with
-// this program.  If not, see <https://www.gnu.org/licenses/>.
-///
-/// Authors: Tony Chen
-
-library;
-
-import 'package:flutter/material.dart';
-
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import 'package:healthpod/features/file/service/components/file_service_widget.dart';
-
-/// The main file management page for HealthPod.
-
-class FileManagementPage extends ConsumerWidget {
-  const FileManagementPage({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return const FileServiceWidget();
-  }
-}
