+++ lib/features/home/service/components/personal_details.dart
@@ -32,6 +32,7 @@ import 'package:flutter/material.dart';
 import 'package:markdown_tooltip/markdown_tooltip.dart';
 import 'package:solidpod/solidpod.dart';
 
+import 'package:healthpod/constants/paths.dart';
 import 'package:healthpod/utils/fetch_profile_data.dart';
 import 'package:healthpod/utils/format_timestamp_for_filename.dart';
 
@@ -239,7 +240,7 @@ class _PersonalDetailsState extends State<PersonalDetails> {
     try {
       // Get all files in the profile directory.
 
-      final dirUrl = await getDirUrl('healthpod/data/profile');
+      final dirUrl = await getDirUrl('$basePath/profile');
       final resources = await getResourcesInContainer(dirUrl);
 
       // Find all profile files.
