+++ lib/home.dart
@@ -27,12 +27,14 @@ library;
 
 import 'package:flutter/material.dart';
 
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:solidpod/solidpod.dart' show getAppNameVersion;
 import 'package:solidui/solidui.dart';
 
+import 'package:healthpod/constants/paths.dart';
 import 'package:healthpod/features/charts/tab.dart';
-import 'package:healthpod/features/file/page.dart';
+import 'package:healthpod/features/file/service/providers/file_service_provider.dart';
 import 'package:healthpod/features/resources/tab.dart';
 import 'package:healthpod/features/table/tab.dart';
 import 'package:healthpod/features/update/tab.dart';
@@ -115,7 +117,7 @@ List<SolidMenuItem> _buildHealthPodMenu() => [
     to **delete** files from your pod storage.
 
     ''',
-        child: const FileManagementPage(),
+        child: const _FileManagementContent(),
       ),
       SolidMenuItem(
         title: 'Support',
@@ -357,6 +359,246 @@ The images for the app were generated by ChatGPT.
         licenses.
 
         ''',
+      ),
+    );
+  }
+}
+
+/// File management content widget using SolidFile.
+
+class _FileManagementContent extends ConsumerStatefulWidget {
+  const _FileManagementContent();
+
+  @override
+  ConsumerState<_FileManagementContent> createState() =>
+      _FileManagementContentState();
+}
+
+class _FileManagementContentState
+    extends ConsumerState<_FileManagementContent> {
+  final GlobalKey<SolidFileBrowserState> _browserKey = GlobalKey();
+
+  @override
+  void initState() {
+    super.initState();
+    // Set up the refresh callback after the widget is built.
+
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      ref.read(fileServiceProvider.notifier).setRefreshCallback(() {
+        _browserKey.currentState?.refreshFiles();
+      });
+      _navigateToFeatureFolder();
+    });
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _navigateToFeatureFolder();
+  }
+
+  /// Navigates to the feature-specific folder.
+
+  void _navigateToFeatureFolder() {
+    final currentPath = ref.read(fileServiceProvider).currentPath ?? basePath;
+    _browserKey.currentState?.navigateToPath(currentPath);
+  }
+
+  /// Creates upload callbacks based on current path.
+
+  SolidFileUploadCallbacks _createUploadCallbacks(String currentPath) {
+    // Determine which type of directory we're in.
+
+    final isInBpDirectory = currentPath.contains('/bp');
+    final isInVaccinationDirectory = currentPath.contains('/vaccination');
+    final isInMedicationDirectory = currentPath.contains('/medication');
+    final isInDiaryDirectory = currentPath.contains('/diary');
+    final isInProfileDirectory = currentPath.contains('/profile');
+
+    // Determine if we should show buttons.
+
+    final showCsvButtons = isInBpDirectory ||
+        isInVaccinationDirectory ||
+        isInMedicationDirectory ||
+        isInDiaryDirectory;
+    final showProfileButtons = isInProfileDirectory;
+
+    return SolidFileUploadCallbacks(
+      onUpload: () => _handleFileUpload(),
+      onImportCsv: showCsvButtons
+          ? () => _handleCsvImport({
+                'isVaccination': isInVaccinationDirectory,
+                'isMedication': isInMedicationDirectory,
+                'isDiary': isInDiaryDirectory,
+                'isBloodPressure': isInBpDirectory,
+              })
+          : null,
+      onExportCsv: showCsvButtons
+          ? () => _handleCsvExport({
+                'isVaccination': isInVaccinationDirectory,
+                'isMedication': isInMedicationDirectory,
+                'isDiary': isInDiaryDirectory,
+                'isBloodPressure': isInBpDirectory,
+              })
+          : null,
+      onImportProfile: showProfileButtons ? () => _handleProfileImport() : null,
+      onExportProfile: showProfileButtons ? () => _handleProfileExport() : null,
+      onVisualiseJson: () => _handleVisualiseJson(),
+      onPreviewFile: () => _handlePreview(),
+      onConvertToJson: () => _handleConvertToJson(),
+    );
+  }
+
+  /// Handles file upload.
+
+  void _handleFileUpload() async {
+    final result = await FilePicker.platform.pickFiles();
+    if (result != null && result.files.isNotEmpty) {
+      final file = result.files.first;
+      if (file.path != null && mounted) {
+        ref.read(fileServiceProvider.notifier).setUploadFile(file.path);
+        await ref.read(fileServiceProvider.notifier).handleUpload(context);
+      }
+    }
+  }
+
+  /// Handles CSV import.
+
+  void _handleCsvImport(Map<String, bool> directoryFlags) {
+    ref.read(fileServiceProvider.notifier).handleCsvImport(
+          context,
+          isVaccination: directoryFlags['isVaccination'] ?? false,
+          isMedication: directoryFlags['isMedication'] ?? false,
+          isDiary: directoryFlags['isDiary'] ?? false,
+          isBloodPressure: directoryFlags['isBloodPressure'] ?? false,
+        );
+  }
+
+  /// Handles CSV export.
+
+  void _handleCsvExport(Map<String, bool> directoryFlags) {
+    ref.read(fileServiceProvider.notifier).handleCsvExport(
+          context,
+          isVaccination: directoryFlags['isVaccination'] ?? false,
+          isDiary: directoryFlags['isDiary'] ?? false,
+          isMedication: directoryFlags['isMedication'] ?? false,
+        );
+  }
+
+  /// Handles Profile import.
+
+  void _handleProfileImport() {
+    debugPrint('Import Profile functionality');
+  }
+
+  /// Handles Profile export.
+
+  void _handleProfileExport() {
+    debugPrint('Export Profile functionality');
+  }
+
+  /// Handles JSON visualisation.
+
+  void _handleVisualiseJson() async {
+    final result = await FilePicker.platform.pickFiles(
+      type: FileType.custom,
+      allowedExtensions: ['json'],
+    );
+    if (result != null && result.files.isNotEmpty) {
+      final file = result.files.first;
+      if (file.path != null) {
+        await _handlePreview();
+      }
+    }
+  }
+
+  /// Handles file preview.
+
+  Future<void> _handlePreview() async {
+    debugPrint('Preview file functionality');
+  }
+
+  /// Handles PDF to JSON conversion.
+
+  void _handleConvertToJson() {
+    debugPrint('Convert to JSON functionality');
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final state = ref.watch(fileServiceProvider);
+    final currentPath = state.currentPath ?? basePath;
+
+    return SolidFile(
+      basePath: basePath,
+      currentPath: currentPath,
+      browserKey: _browserKey,
+      onBackPressed: () {
+        const rootPath = basePath;
+        if (state.currentPath != rootPath) {
+          ref.read(fileServiceProvider.notifier).updateCurrentPath(rootPath);
+          _browserKey.currentState?.navigateToPath(rootPath);
+        }
+      },
+      onFileSelected: (fileName, filePath) {
+        ref.read(fileServiceProvider.notifier)
+          ..setDownloadFile(filePath)
+          ..setFilePreview(fileName)
+          ..setRemoteFileName(fileName);
+      },
+      onFileDownload: (fileName, filePath) async {
+        ref.read(fileServiceProvider.notifier)
+          ..setDownloadFile(filePath)
+          ..setRemoteFileName(fileName)
+          ..handleDownload(context);
+      },
+      onFileDelete: (fileName, filePath) async {
+        // Show confirmation dialogue before deleting.
+
+        final bool? confirm = await showDialog<bool>(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: const Text('Confirm Delete'),
+              content: Text(
+                'Are you sure you want to delete "$fileName"?',
+              ),
+              actions: [
+                TextButton(
+                  onPressed: () => Navigator.of(context).pop(false),
+                  child: const Text('Cancel'),
+                ),
+                TextButton(
+                  onPressed: () => Navigator.of(context).pop(true),
+                  child: const Text('Delete'),
+                ),
+              ],
+            );
+          },
+        );
+
+        if (confirm == true) {
+          // Delete functionality would be implemented here.
+
+          debugPrint('Delete file: $filePath');
+          _browserKey.currentState?.refreshFiles();
+        }
+      },
+      onImportCsv: (fileName, filePath) {
+        // Import CSV functionality would be implemented here.
+
+        debugPrint('Import CSV: $fileName at $filePath');
+      },
+      onDirectoryChanged: (path) {
+        ref.read(fileServiceProvider.notifier).updateCurrentPath(path);
+      },
+      uploadCallbacks: _createUploadCallbacks(currentPath),
+      uploadState: SolidFileUploadState(
+        uploadInProgress: state.uploadInProgress,
+        importInProgress: state.importInProgress,
+        exportInProgress: state.exportInProgress,
+        uploadedFilePath: state.uploadFile,
+        uploadDone: state.uploadDone,
       ),
     );
   }
